Renamed auth > utils. Added test to make sure the User model is the custom one. Added utils to settings.
